api-talkpacific/pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "api-talkpacific"
version = "0.1.0"
description = "API for TalkPacific"
authors = ["Kyle Madsen <kyle.mdsn@gmail.com>"]
readme = "README.md"
packages = [{include = "api_talkpacific"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
langchain = "^0.2.11"
langchain-openai = "^0.1.19"
langchain-community = "^0.2.10"
pydantic = "^2.8.2"
pymongo = { version = "^4.8.0", extras = ["srv"] }
setuptools = "^72.1.0"
uvicorn = "^0.30.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = ">=6.0"
mongomock = "^4.1.2"
flake8 = ">=7.1.0"
toml = "^0.10.2"
python-dotenv = "^1.0.1"
